refactored player component
