+++ src/_components/AudioPlayer/TrackInfo/index.ts
@@ -1,0 +1,1 @@
+export { TrackInfo } from './TrackInfo';
