+++ src/_components/AudioPlayer/VolumeControl/index.ts
@@ -1,0 +1,1 @@
+export { VolumeControl } from './VolumeControl';
