+++ src/_components/AudioPlayer/index.ts
@@ -1,0 +1,1 @@
+export { AudioPlayer } from './AudioPlayer';
