+++ src/_utils/api/index.ts
@@ -1,0 +1,6 @@
+export const fetcher = (url: string) =>
+  fetch(url, {
+    headers: {
+      'Content-Type': 'json',
+    },
+  }).then((res) => res.json());
