+++ src/app/styles.module.css
@@ -40,23 +40,6 @@
   }
 }
 
-.player {
-  display: flex;
-  max-width: 500px;
-  width: 100%;
-  border-radius: 24px;
-  background-color: #202020;
-  /* border: 3px solid #c2c2c2; */
-  padding: 35px;
-}
-
-@media (max-width: 1023px) {
-  .player {
-    padding: 16px;
-    border-radius: 12px;
-  }
-}
-
 .content {
   border-radius: 24px;
   background-color: #707070;
@@ -77,99 +60,6 @@
   height: 525px;
   border-radius: 7px;
   background-color: #c9c9c9;
-}
-
-.controlWrapper {
-  flex: 0 0 100%;
-  max-width: 100%;
-}
-
-.buttons {
-  max-width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 35px;
-}
-
-@media (max-width: 1023px) {
-  .buttons {
-    margin-bottom: 24px;
-  }
-}
-
-.slideContainer {
-  display: flex;
-  align-items: center;
-  height: 8px;
-  position: relative;
-  width: 50%;
-}
-
-.slideStatusContainer {
-  width: 100%;
-  z-index: 9;
-  position: absolute;
-  height: 8px;
-  border-radius: 6px;
-  top: 50%;
-  left: 0;
-  transform: translateY(-50%);
-  overflow: hidden;
-}
-
-.slideStatus {
-  height: 8px;
-  background: #ff7477;
-}
-
-.volume {
-  display: flex;
-  justify-content: center;
-}
-
-.volumeSlide {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 100%;
-  height: 8px;
-  border-radius: 6px;
-  background: #d3d3d3;
-  outline: none;
-}
-
-.volumeSlide::-webkit-slider-thumb {
-  position: relative;
-  z-index: 10;
-  -webkit-appearance: none;
-  appearance: none;
-  width: 21px;
-  height: 21px;
-  background: #ff7477;
-  cursor: pointer;
-  border-radius: 50%;
-}
-
-.volumeSlide::-moz-range-thumb {
-  position: relative;
-  z-index: 10;
-  width: 21px;
-  height: 21px;
-  background: #ff7477;
-  cursor: pointer;
-  border-radius: 50%;
-}
-
-@media (max-width: 1023px) {
-  .volumeSlide::-webkit-slider-thumb {
-    width: 18px;
-    height: 18px;
-  }
-
-  .volumeSlide::-moz-range-thumb {
-    width: 18px;
-    height: 18px;
-  }
 }
 
 .navList {
@@ -205,31 +95,6 @@
   margin-right: 8px;
 }
 
-.control {
-  flex-shrink: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 48px;
-  height: 48px;
-  border-radius: 9px;
-  background-color: #f57073;
-  transition: 0.15s background-color ease;
-  color: #2a2a2a;
-  margin-left: 16px;
-  outline: none;
-  border: none;
-  box-shadow: none;
-}
-
-.control:hover {
-  cursor: pointer;
-}
-
-.control:not(:last-child) {
-  margin-right: 25px;
-}
-
 .socials {
   margin-top: 50px;
   width: 100%;
@@ -249,54 +114,4 @@
 
 .socialItem:not(:last-of-type) {
   margin-right: 12px;
-}
-
-.trackInfo {
-  display: flex;
-  align-items: center;
-  max-width: calc(100% - 48px - 16px);
-}
-
-.trackText {
-  max-width: calc(100% - 48px - 16px);
-}
-
-.trackArtist {
-  font-size: 18px;
-  font-weight: 700;
-  margin: 0 0 4px;
-}
-
-@media (max-width: 1023px) {
-  .trackArtist {
-    font-size: 16px;
-  }
-}
-
-.trackTitle {
-  font-size: 14px;
-  line-height: 1;
-  max-width: 100%;
-  text-wrap: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-@media (max-width: 1023px) {
-  .trackTitle {
-    font-size: 12px;
-  }
-}
-
-.trackCover {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  width: 48px;
-  height: 48px;
-  border-radius: 4px;
-  background-color: #111;
-  overflow: hidden;
-  margin-right: 16px;
 }
